XMLHttpRequest/send-response-event-order.htm

<!DOCTYPE html>
<html>
<head>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" />
    <script src="/resources/testharness.js"></script>
    <title>XMLHttpRequest L2: The send() method: event order when synchronous flag is unset</title>
</head>

<body>
    <div id="log"></div>

    <script type="text/javascript">
        var test = async_test();

        test.step(function() 
        {
            var xhr = new XMLHttpRequest();
            var expect =  ["loadstart", "upload.loadstart", "upload.progress", "upload.load", "upload.loadend", "progress", 4, "load", "loadend"];
            var actual = [];

            xhr.onreadystatechange = function() 
            {
                test.step(function() 
                {
                    if (xhr.readyState == 4) 
                    {
                        actual.push(xhr.readyState);
                    }
                });
            };

            xhr.onloadstart        = function(e){ actual.push(e.type); };
            xhr.onload             = function(e){ actual.push(e.type); };
            xhr.onloadend          = function(e){ actual.push(e.type); VerifyResult()};
            xhr.onprogress         = function(e){ actual.push(e.type);};
            
            xhr.upload.onloadstart = function(e){ actual.push("upload." + e.type); };
            xhr.upload.onload      = function(e){ actual.push("upload." + e.type); };
            xhr.upload.onloadend   = function(e){ actual.push("upload." + e.type);};
            xhr.upload.onprogress  = function(e){ actual.push("upload." + e.type);};

            function VerifyResult()
            {
                test.step(function()
                {
                    assert_array_equals(actual, expect);
                    test.done();
                });
            };

            xhr.open("POST", "./support/echo.php", true);
            xhr.send("Test Message");
        });
    </script>
</body>
</html>
