LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/send-response-event-order-expected.txt


FAIL XMLHttpRequest: The send() method: event order when synchronous flag is unset assert_equals: expected "upload.loadstart(0,12,true)" but got "upload.loadstart(0,0,false)"
